refactor(wlroots): move protocol code
