+++ source/MaaWlRootsControlUnit/CMakeLists.txt
@@ -1,4 +1,4 @@
-file(GLOB_RECURSE maa_wlroots_control_unit_src *.h *.hpp *.cpp)
+file(GLOB_RECURSE maa_wlroots_control_unit_src *.h *.hpp *.cpp Protocol/*.c)
 file(GLOB_RECURSE maa_wlroots_control_unit_header ${MAA_PUBLIC_INC}/MaaControlUnit/WlRootsControlUnitAPI.h ${MAA_PUBLIC_INC}/MaaControlUnit/ControlUnitAPI.h)
 
 add_library(MaaWlRootsControlUnit SHARED ${maa_wlroots_control_unit_src} ${maa_wlroots_control_unit_header})
