--- source/MaaWlRootsControlUnit/Client/WaylandProtocol.h
@@ -1,8 +1,0 @@
-#pragma once
-
-extern "C"
-{
-#include "wayland/virtual-keyboard-priv.h"
-#include "wayland/wlr-screencopy-priv.h"
-#include "wayland/wlr-virtual-pointer-priv.h"
-}
